Generalized at best Scenarios
Tried to generalize scenarios to facilitate props type additions without adding too much classes or having to create a scriptable object class for each case.

Settled for generalisation of all props to inherit scriptable object class Props which is now used in Node.
Use a specific data such as Dialog require a casting to the wanted props type. See the attribute type from Props to see if the casting is possible.
For the serialization, a global props read model is now in use which contains data from all types of props. The conversion to a typed props is done by a PropsFactory which returns a Props (though typed data still remain in wait for casting).